update to code

+++ NotesInteraction/notes.go
@@ -4,6 +4,7 @@ import (
 	"bufio"
 	"fmt"
 	c "notes/Constants"
+	ui "notes/UserInteraction"
 	"os"
 	"strconv"
 	"strings"
@@ -23,8 +24,8 @@ func GetNotes(path string) []string {
 	}
 	return lines
 }
-func ShowNotes(path string) {
 
+func ShowNotes(path string) {
 	// call already made function to get the notes. 😊
 	notes := GetNotes(path)
 	if len(notes) == 0 { // check length of string[] to determine if there are notes.
@@ -36,6 +37,7 @@ func ShowNotes(path string) {
 		fmt.Printf(c.Blue+"%03d - %s\n"+c.Reset, index+1, line)
 	}
 }
+
 func AddNotes(path string) {
 	file, _ := os.OpenFile(path, os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0644)
 	defer file.Close()
@@ -57,8 +59,8 @@ func AddNotes(path string) {
 		}
 	}
 }
+
 func DeleteNotes(path string) {
-	var x string
 	var lineToDelete int    // get slice of lines
 	lines := GetNotes(path) // Get the index to delete
 
@@ -67,34 +69,39 @@ func DeleteNotes(path string) {
 		return
 	}
 	for {
-		fmt.Printf(c.Blue+c.Bold+"\n"+c.DELETE_NOTE_NUM+c.ENTER_VALID_NUM+"%d to %d"+c.TO_CANCEL+"\n"+c.Reset, 1, len(lines))
-		fmt.Scanln(&x)
+		fmt.Printf(c.Blue+c.Bold+"\n"+c.INDICATE_REMOVAL+c.ENTER_VALID_NUM+"%d to %d. "+c.TO_CANCEL+"\n"+c.Reset, 1, len(lines))
+		fmt.Print(c.Blue + c.Bold + "\n" + c.SELECT_OP + c.Reset)
 
-		lineToDelete, _ = strconv.Atoi(x)
+		del := ui.GetInput()
+
+		lineToDelete, _ = strconv.Atoi(del)
 		if lineToDelete == 0 { // Exits if user wants to cancel
 			return
 		}
 		if lineToDelete <= 0 || lineToDelete > len(lines) {
 			if len(lines) > 0 {
-				fmt.Printf(c.Red+c.Bold+c.Italic+c.MAX_NUM+"%d\n"+c.Reset, len(lines))
+				fmt.Printf(c.Red+c.Bold+c.Italic+"\n"+c.MAX_NUM+"%d\n"+c.Reset, len(lines))
 			}
 			continue
 		}
-		lineToDelete--                                                  // shifting to zero index
-		lines = append(lines[:lineToDelete], lines[lineToDelete+1:]...) //appends notes before and after the one to delete
-		//if lineToDelete > 0 || lineToDelete <= len(lines) { //number of note to delete within the existing amount
-		output := ""                 //rebuilds the output string
-		for _, line := range lines { //Loops through the remaining lines,
-			output += line + "\n" //concatenating them into the output string, adding a newline after each line.
+		lineToDelete--                                     // shifting to zero index
+		if lineToDelete > 0 || lineToDelete < len(lines) { //number of note to delete within the existing amount
+			fmt.Printf(c.Blue+"\n"+c.DELETED_NOTE+"%d\n"+c.Reset, lineToDelete+1) //added indication which note was removed
+
+			lines = append(lines[:lineToDelete], lines[lineToDelete+1:]...) //appends notes before and after the one to delete
+
+			output := ""                 //rebuilds the output string
+			for _, line := range lines { //Loops through the remaining lines,
+				output += line + "\n" //concatenating them into the output string, adding a newline after each line
+			}
+			if len(output) > 0 { // Removes the last newline if output is not empty
+				output = output[:len(output)-1]
+			}
+			err := os.WriteFile(path, []byte(output), 0644)
+			if err != nil {
+				panic(err)
+			}
+			break
 		}
-		if len(output) > 0 { // Removes the last newline if output is not empty
-			output = output[:len(output)-1]
-			//	fmt.Println(c.Blue + "Note deleted." + c.Reset)
-		}
-		err := os.WriteFile(path, []byte(output), 0644)
-		if err != nil {
-			panic(err)
-		}
-		break
 	}
 }
